[homewizard] Fix `NumberFormatException` (#17792)

+++ bundles/org.openhab.binding.homewizard/src/main/java/org/openhab/binding/homewizard/internal/dto/DataPayload.java
@@ -58,13 +58,13 @@ public class DataPayload {
     private int longPowerFailCount;
 
     @SerializedName("active_voltage_v")
-    private int activeVoltage;
+    private double activeVoltage;
     @SerializedName("active_voltage_l1_v")
-    private int activeVoltageL1;
+    private double activeVoltageL1;
     @SerializedName("active_voltage_l2_v")
-    private int activeVoltageL2;
+    private double activeVoltageL2;
     @SerializedName("active_voltage_l3_v")
-    private int activeVoltageL3;
+    private double activeVoltageL3;
 
     @SerializedName("active_current_a")
     private double activeCurrent;
@@ -175,7 +175,7 @@ public class DataPayload {
      *
      * @return current active voltage
      */
-    public int getActiveVoltage() {
+    public double getActiveVoltage() {
         return activeVoltage;
     }
 
@@ -184,7 +184,7 @@ public class DataPayload {
      *
      * @return active voltage on phase 1
      */
-    public int getActiveVoltageL1() {
+    public double getActiveVoltageL1() {
         return activeVoltageL1;
     }
 
@@ -193,7 +193,7 @@ public class DataPayload {
      *
      * @return active voltage on phase 2
      */
-    public int getActiveVoltageL2() {
+    public double getActiveVoltageL2() {
         return activeVoltageL2;
     }
 
@@ -202,7 +202,7 @@ public class DataPayload {
      *
      * @return active voltage on phase 3
      */
-    public int getActiveVoltageL3() {
+    public double getActiveVoltageL3() {
         return activeVoltageL3;
     }
